Extract make_shm_pool
