+++ examples/client/make_shm_pool.c
@@ -1,0 +1,45 @@
+/*
+ * Copyright © Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2 or 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "make_shm_pool.h"
+
+#include <fcntl.h>
+#include <unistd.h>
+#include <sys/mman.h>
+
+struct wl_shm_pool*
+make_shm_pool(struct wl_shm* shm, int size, void **data)
+{
+    int fd = memfd_create("make_shm_pool", MFD_CLOEXEC);
+
+    if (fd < 0) {
+        return NULL;
+    }
+
+    posix_fallocate(fd, 0, size);
+
+    *data = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+    if (*data == MAP_FAILED) {
+        close(fd);
+        return NULL;
+    }
+
+    struct wl_shm_pool *pool = wl_shm_create_pool(shm, fd, size);
+
+    close(fd);
+
+    return pool;
+}
